% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf_utils.R
\name{as.jobj}
\alias{as.jobj}
\title{Retrieve an underlying Java object}
\usage{
as.jobj(x)
}
\arguments{
\item{x}{The input, which is some R object encapsulating a Java oject}
}
\description{
Retrieve the underlying Java object from the input
}
\examples{

library(sparklyr)
library(sparklyr.flint)

sc <- try_spark_connect(master = "local")

if (!is.null(sc)) {
  sdf <- copy_to(sc, tibble::tibble(t = seq(10), v = seq(10)))
  ts <- from_sdf(sdf, is_sorted = TRUE, time_unit = "SECONDS", time_column = "t")
  print(ts \%>\% as.jobj)
} else {
  message("Unable to establish a Spark connection!")
}

}
