% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf_utils.R
\name{fromRDD}
\alias{fromRDD}
\title{Construct a TimeSeriesRDD from a Spark RDD of rows}
\usage{
fromRDD(
  rdd,
  schema,
  is_sorted = FALSE,
  time_unit = .sparklyr.flint.globals$kValidTimeUnits,
  time_column = .sparklyr.flint.globals$kDefaultTimeColumn
)
}
\arguments{
\item{rdd}{A Spark RDD[Row] object containing time series data}

\item{schema}{A Spark StructType object containing schema of the time series
data}

\item{is_sorted}{Whether the rows being imported are already sorted by time}

\item{time_unit}{Time unit of the time column (must be one of the following
values: "NANOSECONDS", "MICROSECONDS", "MILLISECONDS", "SECONDS",
"MINUTES", "HOURS", "DAYS"}

\item{time_column}{Name of the time column}
}
\description{
Construct a TimeSeriesRDD containing time series data from a Spark RDD of rows
}
\examples{
\dontrun{
library(sparklyr)
library(sparklyr.flint)

sc <- spark_connect(master = "local")

sdf <- copy_to(sc, tibble::tibble(t = seq(10), v = seq(10))
rdd <- spark_dataframe(sdf) \%>\% invoke("rdd")
schema <- spark_dataframe(sdf) \%>\% invoke("schema")
ts <- fromRDD(
   rdd, schema, is_sorted = TRUE, time_unit = "SECONDS", time_column = "t"
)
}

}
