% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf_utils.R
\name{fromSDF}
\alias{fromSDF}
\title{Construct a TimeSeriesRDD from a Spark DataFrame}
\usage{
fromSDF(
  sdf,
  is_sorted = FALSE,
  time_unit = .sparklyr.flint.globals$kValidTimeUnits,
  time_column = .sparklyr.flint.globals$kDefaultTimeColumn
)
}
\description{
Construct a TimeSeriesRDD containing time series data from a Spark DataFrame
}
