% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Dependencies and initialization procedures}
\description{
Functions in this file specify all runtime dependencies of sparklyr.flint
and package-wide constants in `.sparklyr.flint.globals`.
}
