% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf_utils.R
\name{jtime_unit}
\alias{jtime_unit}
\title{Converter from time unit name to Java enum value}
\usage{
jtime_unit(sc, time_unit = .sparklyr.flint.globals$kValidTimeUnits)
}
\description{
Convenience function to convert from the name of a time unit (e.g.,
"SECONDS") to its corresponding Java TimeUnit enum value
}
