% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizers.R
\name{new_window_obj}
\alias{new_window_obj}
\title{Evaluate a time window specification and instantiate the corresponding time
window object}
\usage{
new_window_obj(sc, window_expr)
}
\description{
Evaluate a time window specification and instantiate the corresponding time
window object
}
