% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf_utils.R
\name{spark_connection.ts_rdd}
\alias{spark_connection.ts_rdd}
\title{Retrieve Spark connection associated with an R object}
\usage{
\method{spark_connection}{ts_rdd}(x)
}
\description{
See \code{\link[sparklyr:spark_connection]{spark_connection}} for more details.
}
\seealso{
Other Spark dataframe utility functions: 
\code{\link{collect.ts_rdd}()},
\code{\link{from_rdd}()},
\code{\link{from_sdf}()},
\code{\link{spark_dataframe.ts_rdd}()},
\code{\link{spark_jobj.ts_rdd}()},
\code{\link{to_sdf}()},
\code{\link{ts_rdd_builder}()}
}
\concept{Spark dataframe utility functions}
