% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf_utils.R
\name{spark_dataframe.ts_rdd}
\alias{spark_dataframe.ts_rdd}
\title{Retrieve a Spark DataFrame}
\usage{
\method{spark_dataframe}{ts_rdd}(x)
}
\description{
Retrieve a Spark DataFrame from a TimeSeriesRDD object
}
