% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf_utils.R
\name{spark_jobj.ts_rdd}
\alias{spark_jobj.ts_rdd}
\title{Retrieve a Spark JVM Object Reference}
\usage{
\method{spark_jobj}{ts_rdd}(x)
}
\description{
See \code{\link[sparklyr:spark_jobj]{spark_jobj}} for more details.
}
\seealso{
Other Spark dataframe utility functions: 
\code{\link{collect.ts_rdd}()},
\code{\link{from_rdd}()},
\code{\link{from_sdf}()},
\code{\link{spark_connection.ts_rdd}()},
\code{\link{spark_dataframe.ts_rdd}()},
\code{\link{to_sdf}()},
\code{\link{ts_rdd_builder}()}
}
\concept{Spark dataframe utility functions}
