% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizers.R
\name{summarize_corr}
\alias{summarize_corr}
\title{Correlation summarizer}
\usage{
summarize_corr(ts_rdd, columns, key_columns = list())
}
\arguments{
\item{ts_rdd}{Timeseries RDD being summarized}

\item{columns}{A list of column names}

\item{key_columns}{Optional list of columns that will form an equivalence
relation associating each row with the time series it belongs to (i.e., any
2 rows having equal values in those columns will be associated with the
same time series, and any 2 rows having differing values in those columns
are considered to be from 2 separate time series and will therefore be
summarized separately)
By default, `key_colums` is empty and all rows are considered to be part of
a single time series.}
}
\description{
Compute pairwise correations among the list of columns specified and store
results in new columns named with the following pattern:
`<column1>_<column2>_correlation` and `<column1>_<column2>_correlationTStat`,
where column1 and column2 are names of any 2 distinct columns
}
