% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizers.R
\name{summarize_corr2}
\alias{summarize_corr2}
\title{Pairwise correlation summarizer}
\usage{
summarize_corr2(ts_rdd, xcolumns, ycolumns)
}
\arguments{
\item{ts_rdd}{Timeseries RDD being summarized}

\item{xcolumns}{A list of column names}

\item{ycolumns}{A list of column names disjoint from xcolumns}
}
\description{
Compute pairwise correations for all possible pairs of columns such that the
first column of each pair is one of `xcolumns` and the second column of each
pair is one of `ycolumns`, storing results in new columns named with the
following pattern:
`<column1>_<column2>_correlation` and `<column1>_<column2>_correlationTStat`
for each pair of columns (column1, column2)
}
