% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizers.R
\name{summarize_var}
\alias{summarize_var}
\title{Variance summarizer}
\usage{
summarize_var(ts_rdd, window, column)
}
\arguments{
\item{ts_rdd}{Timeseries RDD being summarized}

\item{window}{R expression specifying time windows to be summarized (e.g.,
`in_past("1h")` to summarize data from looking behind 1 hour at each time
 point, `in_future("5s")` to summarize data from looking forward 5 seconds
 at each time point)}

\item{column}{Column to be summarized}
}
\description{
Compute variance of values from `column` within each time window and store
results in a new column named `<column>_variance`, with Bessel's correction
applied to the results
}
