% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizers.R
\name{summarize_weighted_avg}
\alias{summarize_weighted_avg}
\title{Weighted average summarizer}
\usage{
summarize_weighted_avg(ts_rdd, window, value_column, weight_column)
}
\arguments{
\item{ts_rdd}{Timeseries RDD being summarized}

\item{window}{R expression specifying time windows to be summarized (e.g.,
`in_past("1h")` to summarize data from looking behind 1 hour at each time
 point, `in_future("5s")` to summarize data from looking forward 5 seconds
 at each time point)}

\item{value_column}{Column containing data point values}

\item{weight_column}{Column containing relative weights of data points}
}
\description{
Compute moving weighted average, weighted standard deviation, weighted t-
stat, and observation count with the column and weight column specified and
store results in new columns named `<value_column>_<weighted_column>_mean`,
`<value_column>_<weighted_column>_weightedStandardDeviation`,
`<value_column>_<weighted_column>_weightedTStat`, and
`<value_column>_<weighted_column>_observationCount`,
}
