% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizers.R
\name{summarize_weighted_corr}
\alias{summarize_weighted_corr}
\title{Pearson weighted correlation summarizer}
\usage{
summarize_weighted_corr(ts_rdd, xcolumn, ycolumn, weight_column)
}
\arguments{
\item{ts_rdd}{Timeseries RDD being summarized}

\item{xcolumn}{Column representing the first random variable}

\item{ycolumn}{Column representing the second random variable}

\item{weight_column}{Column specifying relative weight of each data point}
}
\description{
Compute Pearson weighted correlation between `xcolumn` and `ycolumn` weighted
by `weight_column` and store result in a new columns named
`<xcolumn>_<ycolumn>_<weight_column>_weightedCorrelation`
}
