% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizers.R
\name{summarize_weighted_covar}
\alias{summarize_weighted_covar}
\title{Weighted covariance summarizer}
\usage{
summarize_weighted_covar(ts_rdd, window, xcolumn, ycolumn, weight_column)
}
\arguments{
\item{ts_rdd}{Timeseries RDD being summarized}

\item{window}{R expression specifying time windows to be summarized (e.g.,
`in_past("1h")` to summarize data from looking behind 1 hour at each time
 point, `in_future("5s")` to summarize data from looking forward 5 seconds
 at each time point)}

\item{xcolumn}{Column representing the first random variable}

\item{ycolumn}{Column representing the second random variable}

\item{weight_column}{Column specifying relative weight of each data point}
}
\description{
Compute unbiased weighted covariance between values from `xcolumn` and
`ycolumn` within each time window, using values from `weight_column` as
relative weights, and store results in a new column named
`"<xcolumn>_<ycolumn>_<weight_column>_weightedCovariance`
}
