% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizers.R
\name{summarizers}
\alias{summarizers}
\alias{summarize_count}
\title{Wrapper functions for commonly used summarizer functions}
\usage{
summarize_count(ts_rdd, window, column = NULL)
}
\arguments{
\item{ts_rdd}{Timeseries RDD being summarized}

\item{window}{Expression specifying time window of the summarization (TODO: example)}

\item{column}{If not NULL, then report the number of non-NULL values in the
column specified within each time window, and store the counts in a new
column named `<column>_count`.
Otherwise the number of rows within each time window is reported, and
stored in a column named `count`.}
}
\description{
R wrapper functions for commonly used Flint summarizer functionalities such as
sum and count.

Count the total number of rows if no column is specified, or the number of
non-null values within the specified column within each time window
}
