% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{try_spark_connect}
\alias{try_spark_connect}
\title{Attempt to establish a Spark connection}
\usage{
try_spark_connect(...)
}
\arguments{
\item{...}{Optional arguments; currently unused.}
}
\value{
a Spark connection object if attempt was successful, or NULL otherwise
}
\description{
Call `sparklyr::spark_connect` with parameters given, returning the
resulting Spark connection object upon success, or NULL upon failure
}
\details{
importFrom sparklyr spark_connect
}
