% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_exprs.R
\name{window_exprs}
\alias{window_exprs}
\alias{in_past}
\alias{in_future}
\title{Time window specifications}
\usage{
in_past(duration, sc)

in_future(duration, sc)
}
\arguments{
\item{duration}{String representing length of the time window containing a
number followed by a time unit (e.g., "10s" or "10sec"), where time unit
must be one of the following: "d", "day", "h", "hour", "min", "minute",
"s", "sec", "second", "ms", "milli", "millisecond", "\deqn{ \mu }s",
"micro", "microsecond", "ns", "nano", "nanosecond"}

\item{sc}{Spark connection (does not need to be specified within the context
of `summarize_*` functions)}
}
\description{
Functions for specifying commonly used types of time windows, which are
commonly used within the context of summarize_* functions (e.g.,
`summarize_count(ts_rdd, in_past("3s"))`) where the Spark connection (`sc`)
is implied to be the same Spark connection the timeseries RDD object has
and does not need to be specified explicitly

Create a sliding time window capuring data within the closed interval of
[current time - duration, current time]

Create a sliding time window capuring data within the closed interval of
[current time, current time + duration]
}
