% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_utils.R
\name{window_utils}
\alias{window_utils}
\alias{in_past}
\alias{in_future}
\title{Functions for instantiating time windows}
\usage{
in_past(sc, duration)

in_future(sc, duration)
}
\arguments{
\item{sc}{Spark connection}

\item{duration}{String representing length of the time window containing a
number followed by a time unit (e.g., "10s" or "10sec"), where time unit
must be one of the following: "d", "day", "h", "hour", "min", "minute",
"s", "sec", "second", "ms", "milli", "millisecond", "${\mu}$s", "micro",
"microsecond", "ns", "nano", "nanosecond"}
}
\description{
Convenience functions for instantiating commonly used types of time windows

Create a sliding time window capuring data within the closed interval of
[current time - duration, current time]

Create a sliding time window capuring data within the closed interval of
[current time, current time + duration]
}
