% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf_utils.R
\name{collect.ts_rdd}
\alias{collect.ts_rdd}
\title{Collect data from a TimeSeriesRDD}
\usage{
\method{collect}{ts_rdd}(x, ...)
}
\arguments{
\item{x}{A com.twosigma.flint.timeseries.TimeSeriesRDD object}

\item{...}{Additional arguments to `sdf_collect()`}
}
\description{
Collect data from a TimeSeriesRDD into a R data frame
}
