% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf_utils.R
\name{fromRDD}
\alias{fromRDD}
\title{Construct a TimeSeriesRDD from a Spark RDD of rows}
\usage{
fromRDD(
  rdd,
  schema,
  is_sorted = FALSE,
  time_unit = .sparklyr.flint.globals$kValidTimeUnits,
  time_column = .sparklyr.flint.globals$kDefaultTimeColumn
)
}
\description{
Construct a TimeSeriesRDD containing time series data from a Spark RDD of rows
}
