% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf_utils.R
\name{sdf_utils}
\alias{sdf_utils}
\title{Utility functions for importing a Spark data frame into a TimeSeriesRDD}
\description{
These functions provide an interface for specifying how a Spark data frame
should be imported into a TimeSeriesRDD (e.g., which column represents time,
whether rows are already ordered by time, and time unit being used, etc)
}
