% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizers.R
\name{summarize_corr}
\alias{summarize_corr}
\title{Correlation summarizer}
\usage{
summarize_corr(ts_rdd, columns)
}
\arguments{
\item{ts_rdd}{Timeseries RDD being summarized}

\item{columns}{A list of column names}
}
\description{
Compute pairwise correations among the list of columns specified and store
results in new columns named with the following pattern:
`<column1>_<column2>_correlation` and `<column1>_<column2>_correlationTStat`,
where column1 and column2 are names of any 2 distinct columns
}
