% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizers.R
\name{summarize_count}
\alias{summarize_count}
\title{Count summarizer}
\usage{
summarize_count(ts_rdd, window, column = NULL)
}
\arguments{
\item{ts_rdd}{Timeseries RDD being summarized}

\item{window}{R expression specifying time windows to be summarized (e.g.,
`in_past("1h")` to summarize data from looking behind 1 hour at each time
 point, `in_future("5s")` to summarize data from looking forward 5 seconds
 at each time point)}

\item{column}{If not NULL, then report the number of values in the column
specified that are not NULL or NaN within each time window, and store the
counts in a new column named `<column>_count`.
Otherwise the number of rows within each time window is reported, and
stored in a column named `count`.}
}
\description{
Count the total number of rows if no column is specified, or the number of
non-null values within the specified column within each time window
}
