% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizers.R
\name{summarize_nth_central_moment}
\alias{summarize_nth_central_moment}
\title{N-th central moment summarizer}
\usage{
summarize_nth_central_moment(ts_rdd, column, n)
}
\arguments{
\item{ts_rdd}{Timeseries RDD being summarized}

\item{column}{Column to be summarized}

\item{n}{The order of moment to calculate}
}
\description{
Computes n-th central moment of the column specified and store result in a
new column named `<column>_<n>thCentralMoment`
}
