% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizers.R
\name{summarize_nth_moment}
\alias{summarize_nth_moment}
\title{N-th moment summarizer}
\usage{
summarize_nth_moment(ts_rdd, column, n)
}
\arguments{
\item{ts_rdd}{Timeseries RDD being summarized}

\item{column}{Column to be summarized}

\item{n}{The order of moment to calculate}
}
\description{
Compute n-th moment of the column specified and store result in a new column
named `<column>_<n>thMoment`
}
