% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizers.R
\name{summarize_z_score}
\alias{summarize_z_score}
\title{Z-score summarizer}
\usage{
summarize_z_score(ts_rdd, column, include_current_observation = FALSE)
}
\arguments{
\item{ts_rdd}{Timeseries RDD being summarized}

\item{column}{Column to be summarized}

\item{include_current_observation}{If true, then use unbiased sample standard
deviation with current observation in z-score calculation, otherwise use
unbiased sample standard deviation excluding current observation}
}
\description{
Computes z-score of the most recent value in the column specified, with
respect to the sample mean and standard deviation observed so far, with the
option for out-of-sample calculation, and store result in a new column named
`<column>_zScore`
}
