% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizers.R
\name{summarizers}
\alias{summarizers}
\alias{summarize_count}
\alias{summarize_sum}
\title{Wrapper functions for commonly used summarizer functions}
\usage{
summarize_count(ts_rdd, window, column = NULL)

summarize_sum(ts_rdd, window, column)
}
\arguments{
\item{ts_rdd}{Timeseries RDD being summarized}

\item{window}{R expression specifying time windows to be summarized (e.g.,
`in_past("1h")` to summarize data from looking behind 1 hour at each time
 point, `in_future("5s")` to summarize data from looking forward 5 seconds
 at each time point)}

\item{column}{Column to perform moving sum operation on}
}
\description{
R wrapper functions for commonly used Flint summarizer functionalities such as
sum and count.

Count the total number of rows if no column is specified, or the number of
non-null values within the specified column within each time window

Compute moving sums on the column specified and store results in a new column
named `<column>_sum`
}
