% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{try_spark_connect}
\alias{try_spark_connect}
\title{Attempt to establish a Spark connection}
\usage{
try_spark_connect(...)
}
\arguments{
\item{...}{Parameters for sparklyr::spark_connect}
}
\value{
a Spark connection object if attempt was successful, or NULL otherwise
}
\description{
Attempt to connect to Apache Spark and return a Spark connection object upon success
}
\examples{

try_spark_connect(master = "local")

}
