% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf_utils.R
\name{ts_rdd_builder}
\alias{ts_rdd_builder}
\title{TimeSeriesRDD builder object}
\usage{
ts_rdd_builder(
  sc,
  is_sorted = FALSE,
  time_unit = .sparklyr.flint.globals$kValidTimeUnits,
  time_column = .sparklyr.flint.globals$kDefaultTimeColumn
)
}
\description{
Builder object containing all required info (i.e., isSorted, timeUnit, and
timeColumn) for importing a Spark data frame into a TimeSeriesRDD
}
