% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_exprs.R
\name{in_past}
\alias{in_past}
\title{Create a sliding time window capturing past and current data}
\usage{
in_past(duration, sc)
}
\description{
Create a sliding time window capuring data within the closed interval of
[current time - duration, current time]
}
